requirements.txt
argon2-cffi
flask
pydantic
pymysql[rsa]
tomli
